Moving some files in the include directory to clean things up
