+++ include/functional_dag/filter_sys.hpp
@@ -13,9 +13,8 @@
  */ 
 #pragma once
 
-#include "functional_dag/dag_impl.hpp"
-#include "functional_dag/dag_node_impl.hpp"
-#include "functional_dag/dag_interface.hpp"
+#include <functional_dag/impl/dag_impl.hpp>
+#include <functional_dag/dag_interface.hpp>
 
 namespace fn_dag {
