+++ include/functional_dag/fn_dag_interface.hpp
@@ -14,6 +14,7 @@
 #pragma once
 
 #include<functional>
+
 #include<functional_dag/dag_interface.hpp>
 
 namespace fn_dag {
